+++ src/Storages/PostgreSQL/PostgreSQLReplicationHandler.cpp
@@ -1,12 +1,12 @@
 #include "PostgreSQLReplicationHandler.h"
 
 #include <Common/setThreadName.h>
-#include <Common/getTableOverride.h>
 #include <Parsers/ASTTableOverrides.h>
 #include <Processors/Transforms/PostgreSQLSource.h>
 #include <Processors/Executors/CompletedPipelineExecutor.h>
 #include <Databases/PostgreSQL/fetchPostgreSQLTableStructure.h>
 #include <Storages/PostgreSQL/StorageMaterializedPostgreSQL.h>
+#include <Interpreters/getTableOverride.h>
 #include <Interpreters/InterpreterDropQuery.h>
 #include <Interpreters/InterpreterInsertQuery.h>
 #include <Interpreters/InterpreterRenameQuery.h>
